backend/requirements.txt
# requirements.txt
flask==3.0.2          # Web framework
python-dotenv==1.0.0  # Environment variables
google-generativeai>=0.3.0  # Gemini API
newspaper3k==0.2.8    # Article scraping
requests==2.31.0      # API calls
beautifulsoup4==4.12.2 # HTML parsing (for custom scraping)
lxml==4.9.4           # XML/HTML parser (required for newspaper3k)